move 'toString' tests to separate file
