+++ types/ramda/tsconfig.json
@@ -220,6 +220,7 @@
         "test/thunkify-tests.ts",
         "test/times-tests.ts",
         "test/toPairs-tests.ts",
-        "test/toPairsIn-tests.ts"
+        "test/toPairsIn-tests.ts",
+        "test/toString-tests.ts"
     ]
 }
